Very primitive island generation code
Generates vertex duplicates and doesn't afraid of anything </meme>.
Oh, and it doesn't randomly generate them yet. I wrote a little
terrain data loader. The files must have the width on the first line,
the height on the second and the data values seperated by commas on
the others. -1 is no terrain.